# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "KAVACH API",
        "description": "Risk, Fraud and Regulatory Intelligence Copilot for Indian Banking",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "List Alerts",
                "description": "List alerts with optional filters",
                "operationId": "list_alerts_api_alerts_get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Severity"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Page Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{alert_id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get Alert Detail",
                "description": "Get detailed information for a specific alert",
                "operationId": "get_alert_detail_api_alerts__alert_id__get",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{alert_id}/evidence": {
            "get": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Get Evidence",
                "description": "Get evidence for an alert",
                "operationId": "get_evidence_api_alerts__alert_id__evidence_get",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EvidenceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Create Evidence",
                "description": "Create evidence pack for an alert",
                "operationId": "create_evidence_api_alerts__alert_id__evidence_post",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EvidenceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{alert_id}/verify": {
            "get": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Verify Evidence",
                "description": "Verify evidence integrity",
                "operationId": "verify_evidence_api_alerts__alert_id__verify_get",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerifyResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{alert_id}/feedback": {
            "post": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Submit Feedback",
                "description": "Submit feedback for an alert",
                "operationId": "submit_feedback_api_alerts__alert_id__feedback_post",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{alert_id}/str-draft": {
            "get": {
                "tags": [
                    "Evidence"
                ],
                "summary": "Get Str Draft",
                "description": "Generate STR draft for an alert",
                "operationId": "get_str_draft_api_alerts__alert_id__str_draft_get",
                "parameters": [
                    {
                        "name": "alert_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Alert Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/STRDraftResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlertDetailResponse": {
                "properties": {
                    "alert": {
                        "$ref": "#/components/schemas/AlertResponse"
                    },
                    "story_en": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Story En"
                    },
                    "story_hi": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Story Hi"
                    },
                    "txn_count": {
                        "type": "integer",
                        "title": "Txn Count"
                    },
                    "total_amount_inr": {
                        "type": "number",
                        "title": "Total Amount Inr"
                    }
                },
                "type": "object",
                "required": [
                    "alert",
                    "story_en",
                    "story_hi",
                    "txn_count",
                    "total_amount_inr"
                ],
                "title": "AlertDetailResponse",
                "description": "Alert detail response model"
            },
            "AlertListResponse": {
                "properties": {
                    "alerts": {
                        "items": {
                            "$ref": "#/components/schemas/AlertResponse"
                        },
                        "type": "array",
                        "title": "Alerts"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "alerts",
                    "total",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "AlertListResponse",
                "description": "Alert list response model"
            },
            "AlertResponse": {
                "properties": {
                    "alert_id": {
                        "type": "string",
                        "title": "Alert Id"
                    },
                    "account_id": {
                        "type": "string",
                        "title": "Account Id"
                    },
                    "customer_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Id"
                    },
                    "typology": {
                        "type": "string",
                        "title": "Typology"
                    },
                    "severity": {
                        "type": "string",
                        "title": "Severity"
                    },
                    "score": {
                        "type": "number",
                        "title": "Score"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "rule_name": {
                        "type": "string",
                        "title": "Rule Name"
                    },
                    "citation": {
                        "type": "string",
                        "title": "Citation"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    },
                    "resolution": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Resolution"
                    },
                    "customer_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Name"
                    },
                    "pan": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pan"
                    }
                },
                "type": "object",
                "required": [
                    "alert_id",
                    "account_id",
                    "customer_id",
                    "typology",
                    "severity",
                    "score",
                    "status",
                    "rule_name",
                    "citation",
                    "created_at"
                ],
                "title": "AlertResponse",
                "description": "Alert response model"
            },
...
            "EvidenceResponse": {
                "properties": {
                    "alert_id": {
                        "type": "string",
                        "title": "Alert Id"
                    },
                    "evidence_json": {
                        "type": "object",
                        "title": "Evidence Json"
                    },
                    "file_path": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "File Path"
                    },
                    "sha256_hash": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sha256 Hash"
                    },
                    "created_by": {
                        "type": "string",
                        "title": "Created By"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    },
                    "presigned_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Presigned Url"
                    }
                },
                "type": "object",
                "required": [
                    "alert_id",
                    "evidence_json",
                    "file_path",
                    "sha256_hash",
                    "created_by",
                    "created_at"
                ],
                "title": "EvidenceResponse",
                "description": "Evidence response model"
            },
            "FeedbackRequest": {
                "properties": {
                    "rating": {
                        "type": "integer",
                        "title": "Rating"
                    },
                    "comment": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment"
                    }
                },
                "type": "object",
                "required": [
                    "rating"
                ],
                "title": "FeedbackRequest",
                "description": "Feedback request model"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "STRDraftResponse": {
                "properties": {
                    "alert_id": {
                        "type": "string",
                        "title": "Alert Id"
                    },
                    "str_draft": {
                        "type": "string",
                        "title": "Str Draft"
                    },
                    "format": {
                        "type": "string",
                        "title": "Format"
                    }
                },
                "type": "object",
                "required": [
                    "alert_id",
                    "str_draft",
                    "format"
                ],
                "title": "STRDraftResponse",
                "description": "STR draft response"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VerifyResponse": {
                "properties": {
                    "verified": {
                        "type": "boolean",
                        "title": "Verified"
                    },
                    "details": {
                        "type": "object",
                        "title": "Details"
                    }
                },
                "type": "object",
                "required": [
                    "verified",
                    "details"
                ],
                "title": "VerifyResponse",
                "description": "Evidence verification response"
            },
...
        }
    }
}